Task 183 (Level 12/Lesson 4/Task 5) - My solutuion

+++ src/com/codegym/task/task10/task1003/Solution.java
@@ -9,7 +9,7 @@ public class Solution {
     public static void main(String[] args) {
         float f = (float) 128.50;
         int i = (int) f;
-        int b = (int) (i + f);
+        int b = (int) ((byte) i + f);
         System.out.println(b);
     }
 }
